refactor: segregate interface used in quiz QUZ-39
