+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/game/GameQuizRepository.kt
@@ -1,0 +1,9 @@
+package com.grepp.quizy.game.domain.game
+
+interface GameQuizRepository {
+
+    fun saveQuiz(gameId: Long, quizId: Long): Long?
+
+    fun findQuizzes(gameId: Long): Set<Long>
+
+}
